Added guide for setting up Nextcloud with FRP
I added a guide with example configuration files on how to setup Nextcloud with FRP (https://github.com/fatedier/frp). I was previously struggling with this exact setup, but got it to work using the help of the community:

https://help.nextcloud.com/t/setup-nextcloud-with-frp-fast-reverse-proxy/214793

Now, I would like to share the config files for FRPC and nginx that worked for me here alongside some explanations on how I set it up. I hope that this could be useful for others trying to achieve the same thing. Any suggestions on how to improve the configs or the explanations are very welcome!

Signed-off-by: Carlos <69718740+xZinio@users.noreply.github.com>